requirements.txt
discord.py>=2.3.2 
python-dotenv>=1.0.0 
requests>=2.31.0 
pypdf>=3.15.0 
python-docx>=0.8.11 
openpyxl>=3.1.2 
python-pptx>=0.6.21 
feedparser>=6.0.10 
tzdata>=2024.1 
groq>=0.9.0 
openai>=1.0.0 
fpdf2>=2.7.0
aiofiles>=0.8.0